% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz_avg.R
\name{fuz.avg}
\alias{fuz.avg}
\title{Averaging trait data at different taxonomic levels and filling missing biological information}
\usage{
fuz.avg(
  txn,
  tab,
  lab,
  txc,
  max.lev = 1,
  genus = "Genus",
  restrict.clade = NULL,
  round.sco = NULL
)
}
\arguments{
\item{txn}{taxonomic tree as a taxa by taxonomic levels data frame, starting from higher levels
on the right (e.g., phylum) to lower levels toward the left (e.g., species).}

\item{tab}{data frame containing fuzzy traits.}

\item{lab}{data frame containing trait and modality identifiers and full labels.}

\item{txc}{taxocenosis, i.e., taxa identified at various levels as a vector.}

\item{max.lev}{column identifier of \code{txn}. It is the highest taxonomic level to which
trait data in \code{txc} can be averaged. Must be element of \code{1:ncol(txn)}.}

\item{genus}{column name of the genus level in \code{txn}.}

\item{restrict.clade}{list of vectors containing taxonomic names. For instance, when averaging
trait data for a given genus, if \code{txn} contains 10 species of this genus and only 2 are
biogeographically relevant (i.e., same province), only those 2 species will be considered.}

\item{round.sco}{if not \code{NULL}, value between 0 and 1 below and above which
the averaged score can be rounded to respectively 0 and 1.}
}
\value{
Returns a list of four objects:
\describe{
\item{\code{tab}}{A data frame containing trait data}
\item{\code{lab}}{A data frame containing trait and modality labels}
\item{\code{tax}}{A data frame containing the taxonomy of \code{txc}}
\item{\code{not.found}}{A vector containing taxa from \code{txc} that could not be documented}
}
}
\description{
As biological trait data are generally defined at the species level, trait data are
often missing in field inventories in which taxa are not always identified at this level.
Besides, sampled species can be absent from trait documentations whereas congeneric homologues
can be documented and used to fill missing information.
}
\examples{
data(MacroTraits)
w <- fuz.tab(which.traits = 8)
tab <- w$tab
lab <- w$lab
### Ophiocten sericeum not documented
### while other species from Ophiocten are documented
txc <- c("Amphiura filiformis", "Corophium",
         "Mysidae", "Isopoda", "Ophiocten sericeum",
         "Unknown")
w <- fuz.avg(txn = MacroTraits$taxonomy[8:13],
             tab = tab, lab = lab, txc = txc)
w$tab
w$tax
w$not.found

}
