% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz_bind.R
\name{fuz.bind}
\alias{fuz.bind}
\title{Binds two or more fuzzy-coded trait data tables}
\usage{
fuz.bind(
  list.tab,
  list.lab,
  column.name = "Column.name",
  traitID = "Trait.ID",
  trait = "Trait",
  modality = "Modality"
)
}
\arguments{
\item{list.tab}{a list of data frames with identical row names, each containing at least one trait.}

\item{list.lab}{a list of trait label data frames that respectively correspond to the data frames in \code{list.tab}.}

\item{column.name}{column names in \code{list.tab} as a variable in \code{list.lab}.}

\item{traitID}{trait identifier in \code{list.lab}.}

\item{trait}{full trait labels in \code{list.lab}.}

\item{modality}{full modality labels in \code{list.lab}.}
}
\value{
Returns a list of two data frames:
\describe{
\item{\code{tab}}{A data frame containing trait data}
\item{\code{lab}}{A data frame containing labels}
}
}
\description{
Merge two or more fuzzy-coded trait data sets and create common trait and modality labels.
Useful when analysing separately different trait data sets next to a global analysis.
}
\examples{
### Building a data frame of traits expressing ecosystem functions
data(MacroTraits)
unique(MacroTraits$labels[c("Trait.ID", "Trait")])

### Biodeposition
w <- fuz.tab(which.traits = c(23, 40))
list.sco <- list(c(0.0005, 0.005, 0.05, 0.5, 5, 15),
                 c(0, 1, 0, 0, 0, 0))
biodep <- fuz.comb(tab = w$tab, lab = w$lab,
                   which.traits = 1:2, list.sco = list.sco)
### biodep clearly exhibits outliers
hist(biodep)
### Convert biodep into a fuzzy trait
biodep.fuz <- fuz.conv(biodep, scan = FALSE,
                       interv.bound = c(0, 0.025, 0.1),
                       row.names = names(biodep))
### Create labels
biodep.fuz <- fuz.lab(tab = biodep.fuz, blo = 4,
                      traits = "Biodeposition",
                      modalities = c("Null", "Low", "Intermediate", "High"))

### Sediment biodiffusive mixing
w <- fuz.tab(which.traits = c(4, 6, 23, 26, 37))
w$lab
list.sco <- list(1:5,                                # Sea floor affinity
                 c(1:5, 0),                          # Substratum depth occupancy
                 c(0.0005, 0.005, 0.05, 0.5, 5, 15), # Body mass
                 0:4,                                # Mobility
                 c(0, 1, 0, 0, 0))                   # Sediment biomixing type
biodif <- fuz.comb(tab = w$tab, lab = w$lab,
                   which.traits = 1:5, list.sco = list.sco)
hist(biodif[biodif > 0])
biodif.fuz <- fuz.conv(biodif, scan = FALSE,
                       interv.bound = c(0, 0.02, 0.04, 0.06),
                       row.names = names(biodif))
### Create labels
biodif.fuz <- fuz.lab(tab = biodif.fuz, blo = 5,
                      traits = "Biodiffusion",
                      modalities = c("Null", "Low", "Intermediate", "High", "Very high"))

### bind both traits
w <- fuz.bind(list(biodep.fuz$tab, biodif.fuz$tab),
              list(biodep.fuz$lab, biodif.fuz$lab))
head(w$tab)
w$lab

}
