% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz_comb.R
\name{fuz.comb}
\alias{fuz.comb}
\title{Creation of synthetic trait}
\usage{
fuz.comb(
  tab,
  lab,
  which.traits = NULL,
  traitID = "Trait.ID",
  list.sco,
  formula = paste("x", 1:length(which.traits), sep = "", collapse = " * "),
  fun = rep("sum", length(which.traits)),
  fuz = rep(TRUE, length(which.traits)),
  sco.scale = TRUE,
  trait.scale = TRUE,
  index.scale = FALSE
)
}
\arguments{
\item{tab}{data frame containing fuzzy-coded traits.}

\item{lab}{data frame containing trait and modality identifiers and full labels.}

\item{which.traits}{integer identifying the traits to be included into a formula.}

\item{traitID}{trait identifier variable in \code{lab}.}

\item{list.sco}{list containing vectors of modality performance scores (one per trait).}

\item{formula}{character string expressing the arithmetic trait combination. Uses \code{x1},..., \code{xn}
that correspond respectively to identifiers in \code{which.traits} for n traits. Cumulative product as default.}

\item{fun}{function combining the scored values within a trait. See details.}

\item{fuz}{logical vector. Which traits should be transformed into frequencies within species profiles?}

\item{sco.scale}{if \code{TRUE}, the modality performance scores are rescaled between 0 and 1}

\item{trait.scale}{if \code{TRUE}, each trait is rescaled between 0 and 1}

\item{index.scale}{if \code{TRUE}, the final synthetic trait is rescaled between 0 and 1}
}
\value{
Returns a synthetic trait as a vector of continuous values
}
\description{
This function enables easy combination of different fuzzy traits into a formula.
Especially useful for trait-based indicator development.
}
\details{
Each trait is a block of trait modality columns, and each modality is attributed
a performance score contained in a respective vector in \code{list.sco}. Then, a unique value per trait
is calculated before being considered in the formula.

Let's take the example of a fuzzy profile 0/1/1 and an associated vector of performance scores 1/2/3. If \code{fuz = TRUE},
the profile becomes 0.0/0.5/0.5. If \code{sco.scale = TRUE}, the performance scores become 0.33/0.67/1.00.
By muliplying both, the species performance profile is 0.000/0.335/0.500. A unique value can be represented
by the weighted mean, i.e., by the total sum = 0.835 indicated by \code{fun}.

If \code{trait.scale = TRUE}, all traits are rescaled between 0 and 1 before being considered in the formula,
which gives them an equal weight. Finally, if \code{index.scale = TRUE}, the final synthetic trait
is rescaled between 0 and 1.
}
\examples{
### Example with biodeposition
### as the product of body mass and suspension feeding
data(MacroTraits)
unique(MacroTraits$labels[c("Trait.ID", "Trait")])
w <- fuz.tab(which.traits = c(23, 40))
w$lab
### Define performance scores for each trait
list.sco <- list(c(0.0005, 0.005, 0.05, 0.5, 5, 15),
                 c(0, 1, 0, 0, 0, 0))
tr <- fuz.comb(tab = w$tab, lab = w$lab,
                 which.traits = 1:2, list.sco = list.sco)
### tr clearly exhibits outliers
hist(tr)
### Convert tr into a fuzzy trait
hist(tr[tr > 0])
tr.fuz <- fuz.conv(tr, scan = FALSE,
                   interv.bound = c(0, 0.025, 0.1),
                   row.names = rownames(w$tab),
                   col.names = c("Null", "Low", "Intermediate", "High"))
head(tr.fuz)
apply(tr.fuz, 2, sum)

}
