% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz_conv.R
\name{fuz.conv}
\alias{fuz.conv}
\title{Convert a continuous trait into a fuzzy one}
\usage{
fuz.conv(
  x,
  scan = TRUE,
  interv.bound = NULL,
  include.0 = FALSE,
  col.names = NULL,
  row.names = NULL
)
}
\arguments{
\item{x}{a continuous trait.}

\item{scan}{if \code{TRUE}, plot \code{hist(x)} and ask for boundary choice.}

\item{interv.bound}{interval boundaries within \code{x}. See details.}

\item{include.0}{if TRUE, plot \code{hist(x)}, otherwise \code{hist(x[x > 0])}.}

\item{col.names}{vector of modality column names for the resulting fuzzy-coded trait.}

\item{row.names}{vector of taxon names.}
}
\value{
Returns a complete disjunctive table, i.e., a block of dummy variables.
}
\description{
\code{fuz.conv} generates a series of binary variables from a continuous trait. The user is free
to define interval boundaries. Useful for technical purposes such as homogenising a data set,
coping with outliers or circumventing problems of multivariate non-linearity.
}
\details{
The function uses \code{cut(x, right = TRUE,...)} in which intervals are closed on the right.
Hence, if 0 should be considered as a separate category, use \code{interv.bound = c(0,...)}.
Press escape to leave the procedure.
}
\examples{
x <- sample(0:10, 100, replace = TRUE)^2
### Choose yourself
\dontrun{
w <- fuz.conv(x)
w
}
### Predefined boundaries
w <- fuz.conv(x, scan = FALSE, interv.bound = c(0, 20, 40))
w

}
