% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz_lab.R
\name{fuz.lab}
\alias{fuz.lab}
\title{Creation of trait and modality labels}
\usage{
fuz.lab(tab, blo, traits, modalities)
}
\arguments{
\item{tab}{data frame containing fuzzy traits.}

\item{blo}{vector of modality counts. Its length equals the number of traits.}

\item{traits}{vector of character strings as trait labels.}

\item{modalities}{vector of character strings as trait modality labels.}
}
\value{
Returns a list of two data frames:
\describe{
\item{\code{tab}}{A data frame containing trait data}
\item{\code{lab}}{A data frame containing labels}
}
}
\description{
\code{fuz.lab} creates trait and modality labels with associated identifiers that are linked
to an existing fuzzy trait data frame.
}
\examples{
### Creating a fuzzy synthetic trait
data(MacroTraits)
unique(MacroTraits$labels[c("Trait.ID", "Trait")])

### Biodeposition
w <- fuz.tab(which.traits = c(23, 40))
list.sco <- list(c(0.0005, 0.005, 0.05, 0.5, 5, 15),
                 c(0, 1, 0, 0, 0, 0))
biodep <- fuz.comb(tab = w$tab, lab = w$lab,
                   which.traits = 1:2, list.sco = list.sco)
### biodep clearly exhibits outliers
hist(biodep)
### Convert biodep into a fuzzy trait
biodep.fuz <- fuz.conv(biodep, scan = FALSE,
                       interv.bound = c(0, 0.025, 0.1),
                       row.names = names(biodep))
### Create labels
biodep.fuz <- fuz.lab(tab = biodep.fuz, blo = 4,
                      traits = "Biodeposition",
                      modalities = c("Null", "Low", "Intermediate", "High"))
head(biodep.fuz$tab)
biodep.fuz$lab

}
