% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz_modif.R
\name{fuz.modif}
\alias{fuz.modif}
\title{Fuzzy-coded trait modification}
\usage{
fuz.modif(
  tab,
  lab,
  which.trait,
  merge,
  fun = max,
  mod = NULL,
  column.name = "Column.name",
  traitID = "Trait.ID",
  modalityID = "Modality.ID",
  trait = "Trait"
)
}
\arguments{
\item{tab}{data frame containing fuzzy traits.}

\item{lab}{data frame containing trait and modality identifiers and full labels.}

\item{which.trait}{integer identifying the trait to be modified.}

\item{merge}{vector of integers corresponding respectively to the modalities of the trait to be modified.
Modalities that attributed the same value will merge into a same modality in the modified version of the trait.}

\item{fun}{function used when merging scores from different modalities. \code{max} by default.}

\item{mod}{vector of modality labels. Its length must equal \code{max(merge)}.}

\item{column.name}{column names in \code{tab} as a variable in \code{lab}.}

\item{traitID}{trait identifier variable in \code{lab}.}

\item{modalityID}{modality identifier variable in \code{lab}.}

\item{trait}{full trait label variable in \code{lab}.}
}
\value{
Returns a list of two data frames:
\describe{
\item{\code{tab}}{A data frame containing trait data}
\item{\code{lab}}{A data frame containing labels}
}
}
\description{
\code{fuz.modif} reshapes a fuzzy-coded trait within an existing taxa by trait modalities cross table
according to desired changes in number of modalities and associated labels. It simplifies a trait
when some of its modalities are either not relevant or their rarity not important.
}
\examples{
data(MacroTraits)
### Create a trait data cross table
unique(MacroTraits$labels[c("Trait.ID", "Trait")])
w <- fuz.tab(which.traits = c(37, 33, 34, 36))
indic <- w$lab$Trait.ID
### Visualising the table
head(w$tab)[indic \%in\% 1:2]
### Visualising the labels
w$lab[indic \%in\% 1:2,]
### Modify the trait "Endo-bioconstruction type"
w <- fuz.modif(tab = w$tab, lab = w$lab, which.trait = 2,
               merge = c(1, 2, 2, 1, 2, 3, 2, 3),
               mod = c("None", "Blind-ended", "Open-ended"))
indic <- w$lab$Trait.ID
### The new version of the trait
w$lab[indic == 2,]
head(w$tab)[indic \%in\% 1:2]

}
