% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz_plot.R
\name{fuz.plot}
\alias{fuz.plot}
\title{Plotting taxa and modlities of a fuzzy trait along a continuous variable}
\usage{
fuz.plot(
  x,
  y,
  mct = stats::median,
  bar = "pc",
  pc = c(0.25, 0.75),
  horiz = TRUE,
  rev.mod = TRUE,
  yax = TRUE,
  ylim = NULL,
  ylab = "y",
  mod = NULL,
  cex.mod = 1,
  pos.mod = 1,
  pch.pt = 1,
  col.pt = "grey",
  bg.pt = NULL,
  cex.pt = 1,
  pch.mct = 20,
  col.mct = 1,
  bg.mct = NULL,
  cex.mct = 2,
  jitter = TRUE,
  amount = 0.15,
  bar.lwd = 1,
  bar.col = 1,
  return = FALSE
)
}
\arguments{
\item{x}{data frame either as a fuzzy trait or containing factors.}

\item{y}{continuous variable as a vector.}

\item{mct}{measure of central tendency, either \code{median} or \code{mean}.}

\item{bar}{error bar representing trait modality variation. \code{"pc"} for percentile,
\code{"sd"} for standard deviation and \code{"ci"} for confidence interval.}

\item{pc}{if \code{bar = "pc"}, vector of lower and upper percentiles, respectively.}

\item{horiz}{if \code{TRUE}, modality distributions are represented horizontally.}

\item{rev.mod}{if \code{TRUE}, reverse the order of modality representation from top to bottom.}

\item{yax}{if \code{FALSE}, no y-axis is displayed.}

\item{ylim}{range of variable y.}

\item{ylab}{y-axis label.}

\item{mod}{modality labels.}

\item{cex.mod}{character size of modality labels.}

\item{pos.mod}{margin line for positioning modality labels.}

\item{pch.pt}{taxon plotting symbol type.}

\item{col.pt}{taxon plotting symbol colour.}

\item{bg.pt}{taxon plotting symbol background colour.}

\item{cex.pt}{taxon plotting symbol size.}

\item{pch.mct}{\code{mct} plotting symbol type.}

\item{col.mct}{\code{mct} plotting symbol colour.}

\item{bg.mct}{\code{mct} plotting symbol background colour.}

\item{cex.mct}{\code{mct} plotting symbol size.}

\item{jitter}{if \code{TRUE}, add noise orthogonally to y within modality distributions.}

\item{amount}{degree of noise if \code{jitter} is \code{TRUE}.}

\item{bar.lwd}{error bar width.}

\item{bar.col}{error bar colour.}

\item{return}{if \code{TRUE}, modality or factor level coordinates are returned.}
}
\description{
This function enables the graphical representation of a trait through its different modalities
along a continuous variable such as a multivariate axis or any other continuous ecological descriptor
along which tha taxa documented for the trait can be positioned.
}
\examples{
### Illustration of trade-off in reproductive strategy
data(MacroTraits)
unique(MacroTraits$labels[c("Trait.ID", "Trait")])
### Species by reproductive traits cross table
w <- fuz.tab(which.traits = c(10, 14:18))
tab <- w$tab
lab <- w$lab
indic <- lab$Trait.ID
### Fuzzy Correspondence Analysis (FCA) with ade4
fuz <- ade4::prep.fuzzy.var(df = tab, col.blocks = table(indic))
fca <- ade4::dudi.fca(fuz, scannf = FALSE)
par(mfrow = c(3, 2), mar = c(4, 8, 3, 1))
### Representation of traits along the first FCA axis
for(i in 1:max(indic)){
  # Order trait modalities according to their respective first axis score
  x <- tab[indic == i][order(fca$co[,1][indic == i])]
  # Similarly for modality labels
  mod <- lab$Modality[indic == i][order(fca$co[,1][indic == i])]
  # Plotting execution
  fuz.plot(x = x, y = fca$li[,1], ylab = "", mod = mod)
  # Add trait label
  mtext(unique(lab$Trait)[i], cex = 1, adj = 0, line = 1)
}

### When x contains factors
x <- data.frame(x1 = gl(2, 60), x2 = gl(2, 30, 120), x3 = gl(3, 10, 120))
y <- runif(nrow(x))
fuz.plot(x, y)
}
