% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz_sel.R
\name{fuz.sel}
\alias{fuz.sel}
\title{Fuzzy-coded trait removal}
\usage{
fuz.sel(
  tab,
  lab,
  which.traits,
  column.name = "Column.name",
  traitID = "Trait.ID",
  modalityID = "Modality.ID",
  delete.null.col = TRUE,
  delete.block = TRUE,
  ori.col.name = FALSE
)
}
\arguments{
\item{tab}{data frame containing fuzzy-coded traits.}

\item{lab}{data frame containing trait and modality identifiers and full labels.}

\item{which.traits}{integer identifying the traits to be selected.}

\item{column.name}{column names in \code{tab} as a variable in \code{lab}.}

\item{traitID}{trait identifier variable in \code{lab}.}

\item{modalityID}{modality identifier variable in \code{lab}.}

\item{delete.null.col}{if \code{TRUE}, removes null trait modality columns in
the resulting cross table and accordingly adjusts the labels.}

\item{delete.block}{if \code{TRUE}, removes traits having only one modality.}

\item{ori.col.name}{if \code{TRUE}, original column names of \code{tab} and in \code{lab$Column.name} are kept.}
}
\value{
Returns a list of two data frames:
\describe{
\item{\code{tab}}{A data frame containing trait data}
\item{\code{lab}}{A data frame containing labels}
}
}
\description{
\code{fuz.sel} selects fuzzy-coded traits within an existing taxa by trait modalities cross table and adjust
accordingling trait and modality labels. Useful when some traits are not relevant or create noise
within a multidimensional pattern.
}
\examples{
data(MacroTraits)
### Create a trait data cross table
unique(MacroTraits$labels[c("Trait.ID", "Trait")])
w <- fuz.tab(which.traits = c(3:7, 25, 28:29, 40))
w$lab
### Select the traits "Sea floor affinity", "Motility" and "Feeding type"
w <- fuz.sel(w$tab, w$lab, which.traits = c(2, 6, 9))
w$lab

}
