% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuz_tab.R
\name{fuz.tab}
\alias{fuz.tab}
\title{Creation of trait data cross table and associated labels}
\usage{
fuz.tab(
  data,
  lab,
  traitID = "Trait.ID",
  modalityID = "Modality.ID",
  species = "Species",
  trait = "Trait",
  modality = "Modality",
  score = "Score",
  which.traits = NULL,
  mis.dat = TRUE,
  delete.null.col = TRUE,
  check.sp.records = TRUE
)
}
\arguments{
\item{data}{long-format trait data frame. If missing, \code{MacroTraits$traits} as default.}

\item{lab}{data frame containing trait and modality identifiers and full labels. If missing, \code{MacroTraits$labels} as default.}

\item{traitID}{column name of trait identifier in \code{data} and \code{lab}.}

\item{modalityID}{column name of modality identifier in \code{data} and \code{lab}.}

\item{species}{column name of modality identifier in \code{data}.}

\item{trait}{column name of full trait label in \code{lab}.}

\item{modality}{column name of full modality label in \code{lab}.}

\item{score}{column name of trait modality score in \code{data}.}

\item{which.traits}{vector of trait identifiers. All traits by default.}

\item{mis.dat}{if \code{TRUE}, indicates missing scores in \code{data}.}

\item{delete.null.col}{if \code{TRUE}, removes null trait modality columns
in the resulting cross table and accordingly adjusts the labels.}

\item{check.sp.records}{if \code{TRUE}, verifies whether all species
are evenly documented: are there species for which trait modalities are missing?}
}
\value{
Returns a list of two data frames:
\describe{
\item{\code{tab}}{A data frame containing trait data}
\item{\code{lab}}{A data frame containing labels}
}
}
\description{
From a long-format data table, returns a cross table of species by fuzzy-coded trait modalities
and associated trait and modality labels.
}
\examples{
data(MacroTraits)

### Show traits and respective identifiers
unique(MacroTraits$labels[c("Trait.ID", "Trait")])
w <- fuz.tab(which.traits = c(3:7, 25, 28:29, 40))
head(w$tab[1:10])
w$lab[1:10,]

### Fuzzy Correspondence Analysis with ade4
indic <- w$lab$Trait.ID
fuz <- ade4::prep.fuzzy.var(df = w$tab, col.blocks = table(indic))
fca <- ade4::dudi.fca(fuz, scannf = FALSE)
par(mfrow = c(3, 3), mar = rep(0.1, 4))
for(i in 1:max(indic)){
  ade4::s.label(fca$li, cpoint = 0, clab = 0, cgrid = 1.5)
  points(fca$li, col = "grey")
  tr <- unique(w$lab$Trait)[i]
  mod <- w$lab$Modality[indic == i]
  ade4::s.distri(fca$li, fuz[indic == i], cstar = 0, cpoint = 0,
                 lab = mod, clab = 1.5, sub = tr, possub = "topleft",
                 csub = 2, add.p = TRUE)
}

}
