% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ref.R
\name{get.ref}
\alias{get.ref}
\title{Extraction of bibliographical sources in MacroTraits data}
\usage{
get.ref(data, lab, ref, species, which.traits, ref.type = FALSE)
}
\arguments{
\item{data}{long-format trait data frame containing referenced records. If missing, \code{MacroTraits$traits} as default.}

\item{lab}{lab data frame containing trait and modality identifiers and full labels. If missing, \code{MacroTraits$labels} as default.}

\item{ref}{data frame containing reference identifiers and full references. If missing, \code{MacroTraits$references} as default.}

\item{species}{vector of selected species to be documented.}

\item{which.traits}{vector of trait identifiers. All traits by default.}

\item{ref.type}{if \code{TRUE}, the type of reference (book, article, thesis...) is added.}
}
\value{
Returns a list of two data frames:
\describe{
\item{\code{tab}}{A data frame containing trait data}
\item{\code{lab}}{A data frame containing labels}
}
}
\description{
Returns the references of specified combinations of species and traits
}
\examples{
data(MacroTraits)
### Visualise the traits and their respective identifiers
unique(MacroTraits$labels[c("Trait.ID", "Trait")])
### Vector of selected species
sp <- c("Corophium volutator", "Lanice conchilega", "Macoma balthica")
### Query
w <- get.ref(sp = sp, which.traits = c(8, 9, 25, 40))
\dontrun{
View(w)
}

}
